@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Link, useLocation } from "react-router-dom";
 import "../styles/Navbar.css";
-import ReorderIcon from "@material-ui/icons/Reorder";
+import ReorderIcon from "@mui/icons-material/Reorder";
 
 function Navbar() {
   const [expandNavbar, setExpandNavbar] = useState(false);
