@@ -4,7 +4,7 @@ import {
   VerticalTimelineElement,
 } from "react-vertical-timeline-component";
 import "react-vertical-timeline-component/style.min.css";
-import WorkIcon from "@material-ui/icons/Work";
+import WorkIcon from "@mui/icons-material/Work";
 
 function Experience() {
   return (
